Ignore property methods with incorrect num args
Previously we assumed that if a method started with "get" or "is" that
it had 0 arguments.  Likewise we assumed that if a method started with
"set" it had exactly 1 argument.  These unchecked assumptions cause
IllegalStateExceptions later on in execution, and impose unnecessary
constraints on method signatures of JSON-B model classes.

Signed-off-by: Andrew Guibert <andy.guibert@gmail.com>