Fix serializing java.sql.Timestamp
SqlTimestampSerializer missed to override toTemporalAccessor,
leading to ClassCastExceptions in the default implementation in
AbstractDateSerializer. Since it does all formatting after
calling toInstant on the value, we can safely use the same code
path as is used for java.util.Date (despite what is said about
the inheritance relationship between Date and Timestamp in their
API documentation).

Closes #687.